pyproject.toml
[project]
name = "images-to-pallete"
version = "0.1.0"
description = "Extract unique colors from images"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
]
